trying to keep the implementation to a minimum

On-behalf-of: @SAP gergely.brautigam@sap.com
