+++ bindings/go/oci/transformer/doc.go
@@ -1,0 +1,5 @@
+// Package transformer provides OCI artifact transformation capabilities.
+//
+// This package enables extraction and transformation of OCI artifacts with
+// media-type-specific handling for various content types, including Helm charts.
+package transformer
