+++ bindings/go/oci/transformer/testdata/oci-layout.tar.gz
